fix(app): extract Environment interface to avoid fileReplacements circular import (#9)
