+++ apps/hearthly-app/src/environments/environment.interface.ts
@@ -1,0 +1,10 @@
+export interface Environment {
+  production: boolean;
+  keycloak: {
+    issuer: string;
+    clientId: string;
+  };
+  graphql: {
+    uri: string;
+  };
+}
