+++ apps/hearthly-app/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-import type { Environment } from './environment';
+import type { Environment } from './environment.interface';
 
 export const environment: Environment = {
   production: true,
